smartcacao ML/requirements.txt
ultralytics==8.1.25
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
pyyaml==6.0.1
torch==2.0.1
torchvision==0.15.2
requests==2.31.0
matplotlib==3.7.2
tqdm==4.65.0
